Clean up code to build cleanly with -pedantic
Various cleanups around the code which shall not have functional impact:

* added: CMake will fail if CMRX functions are used before CMRX source
  tree is added via add_subdirectory(). If this happens some properties
  are not set.
* fixed: Warnings emitted if compiler was ran with -pedantic option.
* changed: Three ARM-specific modules dealing with ISR handlers were
  merged together: nvic.c, pendsv.c and syscall.c. The ultimate reason
  for this is to put SVC_Handler into module which contains some
  externally referenced functions so it won't be dropped by the linker.
* changed: Syscall definition tables now have "used" attribute to remove
  warnings that they are defined static and unused. They are picked up
  by the linker and compiled into syscall table.
* added: Mute warnings on VLA use by queues. The code is not portable
  but it is otherwise sound.
