+++ src/os/arch/arm/cortex.c
@@ -1,7 +1,10 @@
-/** @defgroup arch_arm_os OS services implementation
- * @ingroup arch_arm
- * @{
- */
+#include <RTE_Components.h>
+#include CMSIS_device_header
+
+#include <cmrx/sys/syscalls.h>
+#include <kernel/syscall.h>
+#include <cmrx/defines.h>
+
 #include <stdint.h>
 #include <kernel/runtime.h>
 #include <kernel/sched.h>
@@ -15,10 +18,30 @@
 #include <kernel/context.h>
 #include <arch/scb.h>
 
-#ifdef KERNEL_HAS_MEMORY_PROTECTION
-#	include <kernel/mpu.h>
-#   include <arch/mpu_priv.h>
-#endif
+#include <kernel/mpu.h>
+#include <arch/mpu_priv.h>
+
+#include <arch/sysenter.h>
+#include <arch/nvic.h>
+
+#include "signal.h"
+
+int os_nvic_enable(uint32_t irq)
+{
+    NVIC_EnableIRQ(irq);
+    return E_OK;
+}
+
+int os_nvic_disable(uint32_t irq)
+{
+    NVIC_DisableIRQ(irq);
+    return E_OK;
+}
+
+/** @defgroup arch_arm_os OS services implementation
+ * @ingroup arch_arm
+ * @{
+ */
 
 void os_request_context_switch()
 {
@@ -47,7 +70,7 @@ __attribute__((naked)) void PendSV_Handler(void)
 	/* Do NOT put anything here. You will clobber context being stored! */
 	cpu_context.old_task->sp = save_context();
 
-    // This assert checks that we are not preempting some other interrupt 
+    // This assert checks that we are not preempting some other interrupt
     // handler. If you assert here, then your interrupt handler priority
     // is messed up. You need to configure PendSV to be the handler with
     // absolutely the lowest priority.
@@ -58,14 +81,14 @@ __attribute__((naked)) void PendSV_Handler(void)
 	struct OS_core_state_t * cpu_state = &core[coreid()];
 
 #ifdef KERNEL_HAS_MEMORY_PROTECTION
-	if (cpu_context.old_parent_process != cpu_context.new_parent_process 
+	if (cpu_context.old_parent_process != cpu_context.new_parent_process
         || cpu_context.old_host_process != cpu_context.new_host_process)
 	{
 //		mpu_store(&old_host_process->mpu, &old_parent_process->mpu);
-		mpu_restore(&cpu_context.new_host_process->mpu, &cpu_context.new_parent_process->mpu);
+		mpu_restore((const MPU_State *) &cpu_context.new_host_process->mpu, (const MPU_State *) &cpu_context.new_parent_process->mpu);
 	}
 #endif
-	
+
 	// Configure stack for incoming process
     // This assumes that all stacks are of same size
 	mpu_set_region(OS_MPU_REGION_STACK, cpu_context.new_stack, sizeof(os_stacks.stacks[0]), MPU_RW);
@@ -96,6 +119,53 @@ __attribute__((naked)) void PendSV_Handler(void)
 	asm volatile(
 			"pop {pc}"
 	);
+}
+
+/** @} */
+
+/** @defgroup arch_arm_syscall System calls implementation
+ * @ingroup arch_arm
+ * @{
+ */
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wcast-function-type"
+
+static SYSCALL_DEFINITION struct Syscall_Entry_t nvic_syscalls[] = {
+	{ SYSCALL_ENABLE_IRQ, (Syscall_Handler_t) &os_nvic_enable },
+	{ SYSCALL_DISABLE_IRQ, (Syscall_Handler_t) &os_nvic_disable },
+};
+
+#pragma GCC diagnostic pop
+
+/** ARM-specific entrypoint for system call handlers.
+ *
+ * This routine is common entrypoint for all syscall routines. It decodes
+ * the syscall ID requested by userspace application and calls generic
+ * method to service the system call. It is callable by executing the SVC
+ * instruction. Code of SVC_Handler will retrieve the requested SVC ID and
+ * let generic machinery to execute specified system call.
+ * @param arg0 syscall argument
+ * @param arg1 syscall argument
+ * @param arg2 syscall argument
+ * @param arg3 syscall argument
+ */
+__attribute__((interrupt)) void SVC_Handler(void)
+{
+	// This assert checks that we are not preempting some other interrupt
+	// handler. If you assert here, then your interrupt handler priority
+	// is messed up. You need to configure PendSV to be the handler with
+	// absolutely the lowest priority.
+	ASSERT(__get_LR() == (void *) 0xFFFFFFFDU);
+
+	uint32_t * psp = (uint32_t *) __get_PSP();
+	sanitize_psp(psp);
+	ExceptionFrame * frame = (ExceptionFrame *) psp;
+	uint16_t * lra = (uint16_t *) frame->pc;
+	uint8_t syscall_id = *(lra - 1);
+    uint32_t rv = os_system_call(frame->r0123[0], frame->r0123[1], frame->r0123[2], frame->r0123[3], syscall_id);
+    *(psp) = rv;
+    return; /*asm volatile("BX lr");*/
 }
 
 /** @} */
